# lab1/docs/swagger.yaml
basePath: /api
definitions:
...
  ds.CreatePvlcMedFormulaRequest:
    properties:
      category:
        type: string
      description:
        type: string
      formula:
        type: string
      gender:
        type: string
      max_age:
        type: integer
      min_age:
        type: integer
      title:
        type: string
    required:
    - category
    - formula
    - gender
    - max_age
    - min_age
    - title
    type: object
...
  ds.PvlcMedFormulaResponse:
    properties:
      category:
        type: string
      description:
        type: string
      formula:
        type: string
      gender:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      max_age:
        type: integer
      min_age:
        type: integer
      title:
        type: string
    type: object
...
  ds.UpdatePvlcMedFormulaRequest:
    properties:
      category:
        type: string
      description:
        type: string
      formula:
        type: string
      gender:
        type: string
      is_active:
        type: boolean
      max_age:
        type: integer
      min_age:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@lungcalc.ru
    name: API Support
    url: http://localhost:8080
  description: |-
    API для расчета должной жизненной емкости легких (ДЖЕЛ)
    Лабораторная работа 4 - Добавление аутентификации и авторизации
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lung Capacity Calculation API
  version: "1.0"
paths:
...
  /api/pvlc-med-formulas:
    get:
      description: Возвращает список формул с возможностью фильтрации
      parameters:
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      - description: Фильтр по полу
        in: query
        name: gender
        type: string
      - description: Минимальный возраст
        in: query
        name: min_age
        type: integer
      - description: Максимальный возраст
        in: query
        name: max_age
        type: integer
      - description: Активные формулы
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.PvlcMedFormulaResponse'
            type: array
      summary: Получение списка формул
      tags:
      - formulas
    post:
      consumes:
      - application/json
      description: Создает новую формулу для расчета ДЖЕЛ (только для модераторов)
      parameters:
      - description: Данные формулы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.CreatePvlcMedFormulaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание новой формулы
      tags:
      - formulas
  /api/pvlc-med-formulas/{id}:
    delete:
      description: Удаляет формулу ДЖЕЛ (только для модераторов)
      parameters:
      - description: ID формулы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление формулы
      tags:
      - formulas
    get:
      description: Возвращает информацию о конкретной формуле ДЖЕЛ
      parameters:
      - description: ID формулы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.PvlcMedFormulaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение конкретной формулы
      tags:
      - formulas
    put:
      consumes:
      - application/json
      description: Обновляет существующую формулу ДЖЕЛ (только для модераторов)
      parameters:
      - description: ID формулы
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.UpdatePvlcMedFormulaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление формулы
      tags:
      - formulas
  /api/pvlc-med-formulas/{id}/add-to-cart:
    post:
      description: Добавляет формулу в заявку-черновик пользователя
      parameters:
      - description: ID формулы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавление формулы в корзину
      tags:
      - formulas
  /api/pvlc-med-formulas/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для формулы ДЖЕЛ в MinIO (только для модераторов)
      parameters:
      - description: ID формулы
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение формулы
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка изображения для формулы
      tags:
      - formulas
securityDefinitions:
  BearerAuth:
    description: 'JWT токен для аутентификации. Формат: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"